+++ ApiImoveis/Models/Locacao.cs
@@ -8,7 +8,7 @@ namespace ApiImoveis.Models
     public partial class Locacao
     {
         public int IdLocacao { get; set; }
-        public int Cpflocatario { get; set; }
+        public long Cpflocatario { get; set; }
         public int IdImovel { get; set; }
     }
 }
