+++ ApiImoveis/Models/Proprietario.cs
@@ -8,7 +8,7 @@ namespace ApiImoveis.Models
     public partial class Proprietario
     {
         public int IdProprietario { get; set; }
-        public int Cpf { get; set; }
+        public long Cpf { get; set; }
         public int IdImovel { get; set; }
     }
 }
